+++ packages/react/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/react
 
+## 3.0.0-beta.118
+
+### Patch Changes
+
+- Updated dependencies [8d9e8ad]
+  - ai@6.0.0-beta.118
+  - @ai-sdk/provider-utils@4.0.0-beta.36
+
+## 3.0.0-beta.117
+
+### Patch Changes
+
+- ai@6.0.0-beta.117
+- @ai-sdk/provider-utils@4.0.0-beta.35
+
 ## 3.0.0-beta.116
 
 ### Patch Changes